Refactor product API endpoints and add brand functionality
This commit refactors the product API endpoints in order to improve code
organization. It adds the ability to create, update, and retrieve brands
in addition to products. The `createBrand()` and `getAllBrand()`
functions are implemented in the product service. The brand-related
routes are authenticated and restricted to users with appropriate
permissions. This change enhances the overall functionality of the API
and allows for more comprehensive product and brand management. No
issues are referenced in this commit.