Refactor user service to handle default address updates more securely
The commit introduces an update to the user service code in order to
improve the handling of default address updates. Previously, it was
possible to set an address to not be default without setting another
address as default first. This commit addresses that issue by throwing
an error when attempting to set an address to not be default without an
alternative default address. The changes ensure that only one address
can be marked as default at a time, preventing any inconsistencies. This
update enhances the reliability and integrity of the user service.
(Issue #123)